Move expand config map converter back out of internal package (#4672)
* Move expand config map converter back out of internal package


* Update CHANGELOG


* Update expand.go

* Update CHANGELOG.md

* Update CHANGELOG.md
