+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 ## 💡 Enhancements 💡
 
+- Expose experimental API `configmapprovider.NewExpandConverter()` (#4672)
 - `service.NewConfigProvider`: copy slice argument, disallow changes from caller to the input slice (#4729)
 - `confighttp` and `configgrpc`: New config option `include_metadata` to persist request metadata/headers in `client.Info.Metadata` (experimental) (#4547)
 - Remove expand cases that cannot happen with config.Map (#4649)
