+++ config/configmapprovider/expand_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package configprovider
+package configmapprovider
 
 import (
 	"context"
@@ -23,7 +23,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/configtest"
 )
 
 func TestNewExpandConverter(t *testing.T) {
@@ -44,12 +43,12 @@ func TestNewExpandConverter(t *testing.T) {
 	t.Setenv("EXTRA_LIST_VALUE_1", valueExtraListElement+"_1")
 	t.Setenv("EXTRA_LIST_VALUE_2", valueExtraListElement+"_2")
 
-	expectedCfgMap, errExpected := configtest.LoadConfigMap(path.Join("testdata", "expand-with-no-env.yaml"))
+	expectedCfgMap, errExpected := loadConfigMap(path.Join("testdata", "expand-with-no-env.yaml"))
 	require.NoError(t, errExpected, "Unable to get expected config")
 
 	for _, test := range testCases {
 		t.Run(test.name, func(t *testing.T) {
-			cfgMap, err := configtest.LoadConfigMap(path.Join("testdata", test.name))
+			cfgMap, err := loadConfigMap(path.Join("testdata", test.name))
 			require.NoError(t, err, "Unable to get config")
 
 			// Test that expanded configs are the same with the simple config with no env vars.
@@ -89,7 +88,7 @@ func TestNewExpandConverter_EscapedEnvVars(t *testing.T) {
 	t.Setenv("MAP_VALUE_2", receiverExtraMapValue)
 
 	// Retrieve the config
-	cfgMap, err := configtest.LoadConfigMap(path.Join("testdata", "expand-escaped-env.yaml"))
+	cfgMap, err := loadConfigMap(path.Join("testdata", "expand-escaped-env.yaml"))
 	require.NoError(t, err, "Unable to get config")
 
 	expectedMap := map[string]interface{}{
@@ -111,4 +110,13 @@ func TestNewExpandConverter_EscapedEnvVars(t *testing.T) {
 		}}
 	require.NoError(t, NewExpandConverter()(context.Background(), cfgMap))
 	assert.Equal(t, expectedMap, cfgMap.ToStringMap())
+}
+
+func loadConfigMap(fileName string) (*config.Map, error) {
+	ret, err := NewFile().Retrieve(context.Background(), "file:"+fileName, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return ret.Get(context.Background())
 }
