+++ config/configmapprovider/testdata/expand-with-all-env.yaml
